Split into unknown parameter/unknown trigger param
